refactor(adapters): extract YamlMcpClientAdapter shared base
Hermes was the only MCP client adapter writing a YAML config document
(a single top-level servers mapping) rather than the JSON mcpServers
schema. Lift its YAML round-trip / sibling preservation / atomic 0o600
write / malformed-file refusal / configure_mcp_server flow into a new
YamlMcpClientAdapter base so a second YAML-backed adapter can reuse it
without duplicating those blocks (the R0801 duplication guard would
otherwise reject the copy).

HermesClientAdapter now only declares its config path, mcp_servers_key,
display name, and the per-server schema transform; behaviour and the
_to_hermes_format static surface are unchanged.
